test(core): split delegation-service.test.ts by primitive
Mirror abort-delegation.test.ts shape: one test file per
primitive. Extract shared fixtures (makeState, makeSteps,
makeSimpleSteps, makeForSteps, makePromptedForSteps) into a
separate module. The root delegation-service.test.ts keeps the
cross-primitive Result-types describe. Total test count unchanged.
